+++ apps/bestofjs-nextjs/src/app/api/og/projects/[slug]/route.tsx
@@ -1,15 +1,16 @@
+/* eslint-disable @next/next/no-img-element */
 import { ImageResponse } from "next/server";
 
 import {
   formatBigNumber,
   getProjectAvatarUrl,
 } from "@/components/core/project-utils";
-import { Box, ImageLayout } from "@/app/api/og/og-utils";
+import { Box, StarIcon, mutedColor } from "@/app/api/og/og-utils";
 import { searchClient } from "@/app/backend";
 
-export const runtime = "edge";
+import { ImageLayout } from "../../og-image-layout";
 
-const mutedColor = "#a1a1aa";
+export const runtime = "edge";
 
 type Context = { params: { slug: string } };
 export async function GET(_req: Request, { params: { slug } }: Context) {
@@ -19,23 +20,14 @@ export async function GET(_req: Request, { params: { slug } }: Context) {
   return new ImageResponse(
     (
       <ImageLayout>
-        <div style={{ display: "flex", alignItems: "center", gap: 64 }}>
+        <Box style={{ alignItems: "center", gap: 64 }}>
           <ProjectLogo project={project} size={200} />
-          <div
-            style={{
-              display: "flex",
-              flex: 1,
-              flexDirection: "column",
-              gap: 32,
-            }}
-          >
-            <div style={{ display: "flex", gap: 32, fontSize: 80 }}>
-              {project.name}
-            </div>
-            <div style={{ color: "#a1a1aa" }}>{project.description}</div>
+          <Box style={{ flex: 1, flexDirection: "column", gap: 32 }}>
+            <Box style={{ gap: 32, fontSize: 80 }}>{project.name}</Box>
+            <div style={{ color: mutedColor }}>{project.description}</div>
             <Trend project={project} />
-          </div>
-        </div>
+          </Box>
+        </Box>
       </ImageLayout>
     ),
     {
@@ -53,38 +45,17 @@ function ProjectLogo({
   size: number;
 }) {
   const imageURL = getProjectAvatarUrl(project, 100, "dark");
-  return <img src={imageURL} width={size} height={size} />;
+  return <img src={imageURL} width={size} height={size} alt={project.name} />;
 }
 
 function ShowStarsTotal({ value }: { value: number }) {
   return (
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "row",
-        alignItems: "center",
-        color: "#a1a1aa",
-      }}
+    <Box
+      style={{ flexDirection: "row", alignItems: "center", color: mutedColor }}
     >
-      <div style={{ display: "flex" }}>{formatBigNumber(value)}</div>
+      <Box>{formatBigNumber(value)}</Box>
       <StarIcon />
-    </div>
-  );
-}
-
-function StarIcon() {
-  return (
-    <svg
-      stroke="currentColor"
-      fill="currentColor"
-      stroke-width="0"
-      viewBox="0 0 24 24"
-      height="1em"
-      width="1em"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path d="M22 9.24l-7.19-.62L12 2 9.19 8.63 2 9.24l5.46 4.73L5.82 21 12 17.27 18.18 21l-1.63-7.03L22 9.24zM12 15.4l-3.76 2.27 1-4.28-3.32-2.88 4.38-.38L12 6.1l1.71 4.04 4.38.38-3.32 2.88 1 4.28L12 15.4z"></path>
-    </svg>
+    </Box>
   );
 }
 
@@ -104,8 +75,4 @@ function Trend({ project }: { project: BestOfJS.Project }) {
       <ShowStarsTotal value={project.stars} />
     </Box>
   ) : null;
-}
-
-function formatDate(date: Date) {
-  return date.toJSON().slice(0, 10);
 }
